pyproject.toml
[project]
name = "qa-guru-advanced-course"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "pytest (==9.0.1)",
    "fastapi (==0.122.0)",
    "email-validator (==2.3.0)",
    "python-dotenv (==1.2.1)",
    "requests (==2.32.5)",
    "pydantic (==2.12.5)",
    "fastapi-pagination (==0.14.3)",
    "uvicorn (==0.38.0)",
    "allure-pytest (==2.15.2)",
    "psycopg2-binary (==2.9.11)",
    "poetry-core (>=2.0.0)",
    "sqlalchemy (==2.0.44)",
    "mimesis (==18.0.0)"
]

[tool.poetry]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
[dependency-groups]
dev = [
    "setuptools (>=80.9.0,<81.0.0)"
]
